`refactor: reorganize project structure and update import paths`
